requirements.txt
# Required installs.. The bot will not work properly without these
discord.py==2.1.0
spotipy==2.21.0
python-dotenv==0.21.0
PyNaCl==1.5.0
wavelink==1.3.3
lyricsgenius==3.0.1
pyyaml==6.0
rich==13.4.2

# Optional installs
# Running pip install -r <this file> will install these aswell..
# Remove them from here if you don't want them
# (You should keep them if you plan on developing the bot)
mypy
pylint
black
isort
